+++ internal/storage/audit_file.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"fmt"
 	"os"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
@@ -67,9 +66,9 @@ func (a *AuditFile) Save(backup Backup) error {
 }
 
 // List all backup information in the storage. As the audit file doesn't store
-// backup extra information, it will be always nil. The backups are ordered by
-// creation date. On error it will return an Error type encapsulated in a
-// traceable error. To retrieve the desired error you can do:
+// backup extra information, it will be always nil. On error it will return an
+// Error type encapsulated in a traceable error. To retrieve the desired error
+// you can do:
 //
 //     type causer interface {
 //       Cause() error
@@ -125,7 +124,7 @@ func (a *AuditFile) List() (Backups, error) {
 			}
 		}
 
-		backups = append(backups, backup)
+		backups.Add(backup)
 	}
 
 	if err := scanner.Err(); err != nil {
@@ -133,7 +132,6 @@ func (a *AuditFile) List() (Backups, error) {
 	}
 
 	a.logger.Infof("storage: backups listed successfully from audit file storage")
-	sort.Sort(backups)
 	return backups, nil
 }